Fix: Add parking_lot dependency and type conversions
- Add parking_lot = "0.12" to Cargo.toml (required by metrics and video modules)
- Fix type mismatches in converter.rs (i32 vs u32 in damage calculations)
- Fix floating point comparison in metrics tests

Fixes compilation errors from CCW session.

+++ src/video/converter.rs
@@ -121,8 +121,8 @@ impl From<DamageRegion> for Rectangle {
         Self {
             left: damage.x as u16,
             top: damage.y as u16,
-            right: (damage.x + damage.width) as u16,
-            bottom: (damage.y + damage.height) as u16,
+            right: ((damage.x as u32) + damage.width) as u16,
+            bottom: ((damage.y as u32) + damage.height) as u16,
         }
     }
 }
